Display error message via snackbar when a parameter change is done and the config server is down (#328)
